requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
bcrypt==5.0.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pymssql==2.3.8
pyodbc==5.2.0
pytest==8.4.1
python-dotenv==1.1.1
python-jose==3.5.0
redis==6.2.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
uvicorn==0.35.0
